Add Create Template command
Right clicking the project in the Solution window will show a
Create Template menu item. This will create a .template.config
directory inside the project directory then create a template.json
file and open it into the text editor. The Create Template menu
is not displayed for non .NET projects nor if the project already
has a template.json file.