Prevent duplicate folder in Solution window
If the template.json file is included in the project the
.template.config folder would be displayed twice.

A remaining problem is that if the template.json file is removed
from the project then the folder is left as a folder item in the
project file which results in just the .template.config folder
being displayed. Ideally the template.json file would be displayed
but faded here.