Register template folder when creating template.json
When the Create Template menu is selected as well as creating a
template.json file the project's directory is registered as a
template folder. This makes the new project template available in
the New Project directory immediately without having to do the
extra step of adding the directory in preferences.