Allow template category to be selected
When creating the template.json file the Template Information dialog
now has a browse button for the temlate category. This opens a
Template Categories dialog where the categories are displayed in a
tree view and the bottom level category can be selected. When the
OK button is clicked the category path is updated in the Template
Information dialog.