Show .templateconfig for solution in Solution window
If the .templateconfig folder exists in the solution's directory
then display it in the Solution window.

Not using a ProjectFolder for the .templateconfig directory in the
solution directory. The AssetFolderNodeBuilderExtension seems to
assume the Project is not null for the ProjectFolder class and
causes a null reference exception when this folder is deleted. So
to workaround this a SolutionTemplateConfigFolder class has been
added.