Prevent templating engine exception being logged many times
When the New Project dialog is opened the custom templates are loaded.
If there was an error then the load would fail multiple times in
quick succession since the New Project dialog calls the
templating provider's GetTemplates method several times. This was
then causing the same exception to be repeatedly logged in the
templating log output window. To avoid this when an error occurs
an attempt is only made to load the templates again after a second
timeout. So when the New Project dialog is opened the exception
will only occur once.