Indicate category is new when creating new project template
The template categories dialog that is displayed when selecting a
template category will now indicate a category is new and requires
a restart before it is available. A check is made against the
custom template category ids and names that are stored in the
template creator's addin.xml file. If there is a mismatch then the
template category is marked as new. The dialog will then show the
template name with "(New category - requires restart)" appended to
the tree item's text.