Support excluding files from being formatted
In the tags section of the template.json file a new property can be
defined vsmac-file-format-exclude or md-file-format-exclude which
can have a semi-colon delimited list of files to exclude from the
file formatting. Limited wildcard support is available, such as
"*.xml" which will do a case insensitive check of the file extension
to see if they end with ".xml". If the file specified does not start
with "*." then the comparison will use the file exclude and see if
the file about to be formatted matches the end of the file exclude.

Fixes #2